toggleButtonを別コンポーネントにする
