+++ src/pages/index.jsx
@@ -4,6 +4,7 @@ import { useContext } from "react";
 import { modeContext } from "./modeProvider";
 
 import { Title } from "./Title";
+import { ToggleModeButton } from "./ToggleModeButton";
 
 const nightMode = css`
   background-color: #472712;
@@ -19,57 +20,11 @@ const image = css`
   height: auto;
 `;
 
-const toggleButton = css`
+const toggleButtonWrap = css`
   display: flex;
   justify-content: right;
   padding-top: 16px;
   padding-right: 16px;
-`;
-
-const input = css`
-  display: none;
-`;
-
-const label = css`
-  position: relative;
-  display: block;
-  width: 64px;
-  height: 32px;
-  border-radius: 16px;
-  background-color: #bdc1ca;
-  &::after {
-    content: "";
-    position: absolute;
-    display: block;
-    width: 24px;
-    height: 24px;
-    transform: translateY(-50%);
-    top: 50%;
-    left: 4px;
-    background-color: white;
-    border-radius: 50%;
-  }
-`;
-
-const labelNightMode = css`
-  position: relative;
-  display: block;
-  width: 64px;
-  height: 32px;
-  border-radius: 16px;
-  background-color: #e8b96b;
-  &::after {
-    content: "";
-    position: absolute;
-    display: block;
-    width: 24px;
-    height: 24px;
-    transform: translateY(-50%);
-    top: 50%;
-    right: 4px;
-    background-color: white;
-    border-radius: 50%;
-  }
 `;
 
 const buttonWrap = css`
@@ -86,27 +41,13 @@ const buttonWrap = css`
   }
 `;
 
-const ToggleModeButton = () => {
-  const { mode, setMode } = useContext(modeContext);
-  const handleClick = () => {
-    setMode(!mode);
-    console.log(mode);
-  };
-  return (
-    <>
-      <input css={input} id="mode" type="checkbox" onChange={handleClick} />
-      <label css={mode ? labelNightMode : label} htmlFor="mode" />
-    </>
-  );
-};
-
 export default function Home() {
   const { mode } = useContext(modeContext);
   const modeStyle = mode ? lightMode : nightMode;
 
   return (
     <main css={modeStyle}>
-      <div css={toggleButton}>
+      <div css={toggleButtonWrap}>
         <ToggleModeButton />
       </div>
       <Title />
